requirements.txt
google-generativeai
SpeechRecognition
pyttsx3
requests
PyAudio


